Decorator Pattern
The ConcreteDecorator is the last point of order. Will likely require some updates in SimpleQuestion.java.

A creative way of storing answers that all of the decorators can have in common is what will be required in order to implement this functionality properly. This will mean a change in SimpleQuestion's constructor.